pyproject.toml
[project]
name = "fa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "sqlmodel>=0.0.24",
    "bcrypt==4.0.1",
]
